backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.4.4
flask-cors==3.0.10
python-dotenv==1.0.0
werkzeug==2.3.6 